Created SchoolContext factory
